+++ packages/webcompy/src/webcompy/components/_hooks.py
@@ -10,7 +10,8 @@ from webcompy.components._libs import Context, generate_id
 from webcompy.di import inject
 from webcompy.di._keys import HYDRATION_DATA_KEY
 from webcompy.hydration._payload import TransferAsyncResultEntry
-from webcompy.signal._base import CallbackConsumerNode, SignalBase
+from webcompy.signal import SignalBase
+from webcompy.signal._base import CallbackConsumerNode
 from webcompy.signal._graph import consumer_destroy
 
 _active_component_context: ContextVar[Context[Any]] = ContextVar("_active_component_context")
